src/tde/pyproject.toml
[project]
name = "tde"
version = "0.1.1"
description = "Technical Debt Extractor - Run dotnet format and output diagnostics in TOON format"
readme = "../../README.md"
authors = [
    { name = "Oleksii Nikiforov", email = "alexey_nikiforovall@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.0",
    "python-toon>=0.1.3",
    "rich>=14.2.0",
]

[project.scripts]
tde = "tde.cli:main"

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"
